refactor(sdk)!: move engine constants to the engine submodule (#1851)
