+++ sdk/packages/node/iii/src/protocol.ts
@@ -1,0 +1,15 @@
+export { MessageType } from './iii-types'
+export type {
+  ErrorBody,
+  RegisterFunctionFormat,
+  RegisterFunctionMessage,
+  RegisterTriggerMessage,
+  RegisterTriggerTypeMessage,
+  TriggerRequest,
+} from './iii-types'
+export type {
+  RegisterFunctionInput,
+  RegisterFunctionOptions,
+  RegisterTriggerInput,
+  RegisterTriggerTypeInput,
+} from './types'
