+++ sdk/packages/node/iii/tsdown.config.ts
@@ -11,6 +11,8 @@ export default defineConfig({
     './src/runtime.ts',
     './src/errors.ts',
     './src/internal.ts',
+    './src/engine.ts',
+    './src/protocol.ts',
   ],
   format: ['esm', 'cjs'],
   dts: true,
